# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RT Backend API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "schemas": {
...
      "Institution": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateInstitutionInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateInstitutionResponse": {
        "type": "object",
        "properties": {
          "institution": {
            "$ref": "#/components/schemas/Institution"
          },
          "fees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Fee"
            }
          }
        },
        "required": [
          "institution",
          "fees"
        ]
      },
      "DeleteInstitutionResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "ok",
          "message"
        ]
      },
      "ReactivateInstitutionResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "ok",
          "message"
        ]
      },
...
      "Student": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "guardianId": {
            "type": "integer"
          },
          "institutionId": {
            "type": "integer"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
      "Fee": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "type": {
            "type": "string"
          },
          "modality": {
            "type": "string",
            "enum": [
              "inPerson",
              "online",
              "cancelled"
            ]
          },
          "numberOfStudents": {
            "type": "integer"
          },
          "guardianAmount": {
            "type": "integer"
          },
          "tutorAmount": {
            "type": "integer"
          },
          "institutionId": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "type",
          "modality",
          "numberOfStudents",
          "guardianAmount",
          "tutorAmount",
          "institutionId",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "AddStudentToGuardianRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Student name (required)"
          },
          "institutionId": {
            "type": "integer",
            "description": "Institution ID (required)"
          },
          "guardianId": {
            "type": "integer",
            "description": "Guardian ID (required)"
          }
        },
        "required": [
          "name",
          "institutionId",
          "guardianId"
        ]
      },
      "AddStudentToGuardianResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "student": {
            "$ref": "#/components/schemas/Student"
          },
          "reactivated": {
            "type": "boolean",
            "description": "True when an existing inactive student was reactivated instead of creating a new record"
          }
        },
        "required": [
          "ok",
          "student"
        ]
      },
      "RemoveStudentFromGuardianRequest": {
        "type": "object",
        "properties": {
          "guardianId": {
            "type": "integer",
            "description": "Guardian ID (required)"
          },
          "studentId": {
            "type": "integer",
            "description": "Student ID (required)"
          }
        },
        "required": [
          "guardianId",
          "studentId"
        ]
      },
      "RemoveStudentFromGuardianResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "ok",
          "message"
        ]
      }
    }
  },
  "paths": {
...
    "/institutions": {
      "post": {
        "summary": "Create an institution",
        "description": "Creates an institution and clones fee rows from institution id=1 (The Grange School).",
        "tags": [
          "Institutions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInstitutionInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created institution",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateInstitutionResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all institutions",
        "tags": [
          "Institutions"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "sendInactive",
            "required": false,
            "schema": {
              "type": "boolean"
            },
            "description": "If false, only active institutions are returned. If true or omitted, returns all."
          }
        ],
        "responses": {
          "200": {
            "description": "List of institutions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Institution"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/institutions/{institutionId}/guardians": {
      "get": {
        "summary": "Get guardians from an institution",
        "tags": [
          "Institutions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "institutionId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Institution ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of guardian"
          }
        }
      }
    },
    "/institutions/search": {
      "get": {
        "summary": "Search institutions by name",
        "tags": [
          "Institutions"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Search query to match institution names"
          }
        ],
        "responses": {
          "200": {
            "description": "List of institutions matching the search query",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Institution"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - search query is required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/institutions/{id}": {
      "delete": {
        "summary": "Delete (deactivate) an institution",
        "description": "Deletes an institution only if there are no pending class payments in the last 12 months and all coordinator payments for those months exist and are completed. Operation is a soft delete that deactivates the institution and its users.",
        "tags": [
          "Institutions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Institution ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Institution deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteInstitutionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Cannot delete due to pending or missing payments",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteInstitutionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/institutions/{id}/reactivate": {
      "patch": {
        "summary": "Reactivate an institution",
        "description": "Reactivates a previously deactivated institution.",
        "tags": [
          "Institutions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Institution ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Institution reactivated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReactivateInstitutionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Institution not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReactivateInstitutionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/students/{guardianId}": {
      "get": {
        "summary": "Get students by guardian ID",
        "description": "Retrieve students associated with a specific guardian.\n- Admins/coordinators: returns all students\n- Guardians/tutors: returns only active students\n",
        "tags": [
          "Students"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "guardianId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Guardian ID"
          },
          {
            "in": "query",
            "name": "sendInactive",
            "required": false,
            "schema": {
              "type": "boolean"
            },
            "description": "If false, only active students are returned for admin/coordinator. If true or omitted, returns all. Guardians/tutors always receive only active students."
          }
        ],
        "responses": {
          "200": {
            "description": "List of students retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "students": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Student"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - user is not authenticated"
          },
          "404": {
            "description": "No students found for the specified guardian"
          }
        }
      }
    },
    "/students/{id}/reactivate": {
      "patch": {
        "summary": "Reactivate a student",
        "description": "Reactivates a previously deactivated student.\n- Requires admin or coordinator role\n",
        "tags": [
          "Students"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Student ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Student reactivated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid student id"
          },
          "403": {
            "description": "Forbidden - user is not an admin or coordinator"
          }
        }
      }
    },
    "/students/add": {
      "post": {
        "summary": "Add a student to a guardian",
        "description": "Create a student for a guardian or reactivate an existing inactive one with the same name (case-insensitive).\n- Requires admin or coordinator role\n- All fields are required\n- If a student with the same name exists for the guardian and is inactive, it will be reactivated instead of creating a new record\n",
        "tags": [
          "Students"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddStudentToGuardianRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Student reactivated or already active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddStudentToGuardianResponse"
                }
              }
            }
          },
          "201": {
            "description": "Student created and added to guardian successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddStudentToGuardianResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or validation error (missing required fields, institution not found)"
          },
          "403": {
            "description": "Forbidden - user is not an admin or coordinator"
          },
          "404": {
            "description": "Guardian not found"
          }
        }
      }
    },
    "/students/delete": {
      "post": {
        "summary": "Remove a student from a guardian",
        "description": "Soft-delete (deactivate) a student by setting isActive = false for the specified guardian/student pair.\n- Requires admin or coordinator role\n- The student must be associated with the specified guardian\n",
        "tags": [
          "Students"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveStudentFromGuardianRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Student removed from guardian successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveStudentFromGuardianResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or validation error (missing required fields)"
          },
          "403": {
            "description": "Forbidden - user is not an admin or coordinator"
          },
          "404": {
            "description": "Student not found for the specified guardian"
          }
        }
      }
    },
...
  },
...
}